Update Delete Category feature

+++ BookStoreWebsite/src/main/java/com/bookstore/service/CategoryServices.java
@@ -113,11 +113,23 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId = Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
+		String message;
 		
+		Category category = categoryDAO.get(categoryId);
+		
+		if (category == null) {
+			message = "Could not find category with ID " + categoryId
+						+ ", or it might have been deleted";
+			request.setAttribute("message", message);
+			request.getRequestDispatcher("message.jsp").forward(request, response);
+			return;
+		} else {	
+			categoryDAO.delete(categoryId);
+			message = "The category " + categoryId + "has been removed successfully.";			
+		}
 		// TODO: Checking for Books belong to a Category
 		// We cannot delete that Category
-		String message = "The category " + categoryId + "has been removed successfully.";
+		
 		listCategory(message);
 	}
 }
